update debugadapter to use interface
