+++ org.eclipse.jdt.ls.debug/src/org/eclipse/jdt/ls/debug/adapter/DebugAdapter.java
@@ -22,7 +22,6 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
 
 import org.eclipse.core.runtime.CoreException;
-import org.eclipse.jdi.Bootstrap;
 import org.eclipse.jdt.ls.debug.DebugEvent;
 import org.eclipse.jdt.ls.debug.DebugUtility;
 import org.eclipse.jdt.ls.debug.IBreakpoint;
@@ -66,6 +65,7 @@ public class DebugAdapter implements IDebugAdapter {
     private BreakpointManager breakpointManager;
     private AtomicInteger nextBreakpointId = new AtomicInteger(1);
     private List<Disposable> eventSubscriptions;
+    private IDebugProvider debugProvider;
     
     private IdCollection<StackFrame> frameCollection = new IdCollection<>();
     private IdCollection<String> sourceCollection = new IdCollection<>();
@@ -74,10 +74,11 @@ public class DebugAdapter implements IDebugAdapter {
     /**
      * Constructor.
      */
-    public DebugAdapter(Consumer<Events.DebugEvent> consumer) {
+    public DebugAdapter(Consumer<Events.DebugEvent> consumer, IDebugProvider provider) {
         this.eventConsumer = consumer;
         this.breakpointManager = new BreakpointManager();
         this.eventSubscriptions = new ArrayList<>();
+        this.debugProvider = provider;
     }
 
     @Override
@@ -255,6 +256,7 @@ public class DebugAdapter implements IDebugAdapter {
         String mainClass = arguments.startupClass;
         String classpath;
         try {
+            // to-do: move to ls
             classpath = AdapterUtils.computeClassPath(arguments.projectName, mainClass);
             classpath = classpath.replaceAll("\\\\", "/");
         } catch (CoreException e) {
@@ -272,7 +274,7 @@ public class DebugAdapter implements IDebugAdapter {
         Logger.logInfo("Launch JVM with main class \"" + mainClass + "\", -classpath \"" + classpath + "\"");
         
         try {
-            this.debugSession = DebugUtility.launch(Bootstrap.virtualMachineManager(), mainClass, classpath);
+            this.debugSession = DebugUtility.launch(debugProvider.getVirtualMachineManager(), mainClass, classpath);
         } catch (IOException | IllegalConnectorArgumentsException | VMStartException e) {
             Logger.logException("Launching debuggee vm exception", e);
             return new Responses.ErrorResponseBody(new Types.Message(3001, "Cannot launch jvm.", null));
@@ -569,7 +571,7 @@ public class DebugAdapter implements IDebugAdapter {
 
     private IBreakpoint[] convertClientBreakpointsToDebugger(String sourceFile, int[] lines) {
         int[] debuggerLines = this.convertClientLineToDebugger(lines);
-        String[] fqns = AdapterUtils.getFullyQualifiedName(sourceFile, debuggerLines);
+        String[] fqns = debugProvider.getFullyQualifiedName(sourceFile, debuggerLines, null);
         IBreakpoint[] breakpoints = new IBreakpoint[lines.length];
         for (int i = 0; i < lines.length; i++) {
             breakpoints[i] = this.debugSession.createBreakpoint(fqns[i], debuggerLines[i]);
@@ -579,7 +581,7 @@ public class DebugAdapter implements IDebugAdapter {
 
     private Types.Source convertDebuggerSourceToClient(Location location) throws URISyntaxException, AbsentInformationException {
         Types.Source source = null;
-        String uri = AdapterUtils.getURI(location.declaringType().name());
+        String uri = debugProvider.getSourceFileURI(location.declaringType().name());
         String name = location.sourceName();
 
         if (uri != null && uri.startsWith("jdt://")) {
@@ -599,7 +601,7 @@ public class DebugAdapter implements IDebugAdapter {
     }
 
     private String convertDebuggerSourceToClient(String uri) {
-        return AdapterUtils.getContents(uri);
+        return debugProvider.getSourceContents(uri);
     }
 
     private Types.Thread convertDebuggerThreadToClient(ThreadReference thread) {
