+++ org.eclipse.jdt.ls.debug/src/org/eclipse/jdt/ls/debug/adapter/IDebugProvider.java
@@ -1,0 +1,4 @@
+package org.eclipse.jdt.ls.debug.adapter;
+
+public interface IDebugProvider extends ISourceLookUpProvider, IVirtualMachineManagerProvider {
+}
