+++ org.eclipse.jdt.ls.debug/src/org/eclipse/jdt/ls/debug/adapter/ProtocolServer.java
@@ -45,7 +45,7 @@ public class ProtocolServer {
      * @param writer
      *              the output writer
      */
-    public ProtocolServer(Reader reader, Writer writer) {
+    public ProtocolServer(Reader reader, Writer writer, IDebugProvider debugProvider) {
         this.reader = reader;
         this.writer = writer;
         this.bodyLength = -1;
@@ -54,7 +54,7 @@ public class ProtocolServer {
         this.eventQueue = new ConcurrentLinkedQueue<>();
         this.debugAdapter = new DebugAdapter(debugEvent -> {
             this.sendEventLater(debugEvent.type, debugEvent);
-        });
+        }, debugProvider);
     }
 
     /**
